Keyframing fix
There was a bug where the wrong variable was referenced when applying keyframe data to range of frames -- variable referencing the index within range of frames was used when updating keyframe data, not the frame at the index. The problem was intermittent, only affecting movies where updating keyframe in earlier frames. This is now referencing the correct variable.